Added a RequestExceptionHandler
