+++ core/src/main/java/com/novoda/merlin/service/HostPinger.java
@@ -19,38 +19,6 @@ class HostPinger {
 
     }
 
-    interface ResponseCodeValidator {
-        boolean isResponseCodeValid(int responseCode);
-        boolean isResponseCodeValid(RequestException e);
-
-        ResponseCodeValidator DEFAULT_ENDPOINT_VALIDATOR = new ResponseCodeValidator() {
-            @Override
-            public boolean isResponseCodeValid(int responseCode) {
-                return responseCode == 204;
-            }
-
-            @Override
-            public boolean isResponseCodeValid(RequestException e) {
-                return false;
-            }
-        };
-
-        ResponseCodeValidator CUSTOM_ENDPOINT_VALIDATOR = new ResponseCodeValidator() {
-            @Override
-            public boolean isResponseCodeValid(int responseCode) {
-                return true;
-            }
-
-            @Override
-            public boolean isResponseCodeValid(RequestException e) {
-                if (e.causedByIO()) {
-                    return false;
-                }
-                throw e;
-            }
-        };
-    }
-
     public static HostPinger newInstance(PingerCallback pingerCallback) {
         MerlinLog.d("Host address not set, using Merlin default : " + Merlin.DEFAULT_ENDPOINT);
         return newInstance(pingerCallback, Merlin.DEFAULT_ENDPOINT);
